services/auth/pyproject.toml
[project]
name = "shopstack-auth"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.6",
  "sqlalchemy>=2.0",
  "asyncpg>=0.29",
  "alembic>=1.13",
  "python-jose[cryptography]>=3.3",
  "passlib[bcrypt]>=1.7",
  "redis>=5.0",
  "httpx>=0.27",
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
]

[tool.uv]
dev-dependencies = ["ruff>=0.6"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
